Create pdfminer_multiprocess_timed.py
Multiprocessing PDFs to HTML conversion (takes a while but works!). The code grabs all the PDFs in a repository you put in. For each document, conversion process is initiated. After 90 seconds, the process is aborted (feel free to adjust timing). Therefore, if the file takes a while to convert (due to horizontal paging or other reasons), the name of the file is printed and the process for that specific file is aborted. The code moves on to the next document and you don't have to manually interrupt it.